Чистка интерфейсов IEdge, IVertex
