+++ GraphLight.Core/Graph/VertexExtensions.cs
@@ -1,0 +1,31 @@
+﻿using System.Linq;
+
+namespace GraphLight.Graph
+{
+    public static class VertexExtensions
+    {
+        public static void Update(this IVertex vertex)
+        {
+            foreach (var e in vertex.Edges.Cast<IEdge>())
+            {
+                var pts = e.Points;
+                using (e.DeferRefresh())
+                {
+                    if (pts.Count == 2 || e.Src == e.Dst)
+                    {
+                        e.UpdateSrcPort();
+                        e.UpdateDstPort();
+                    }
+                    else if (e.Src == vertex)
+                        e.UpdateSrcPort();
+                    else
+                        e.UpdateDstPort();
+                    var first = e.Points.First();
+                    var last = e.Points.Last();
+                    e.Data.FixDraggablePoints(first);
+                    e.Data.FixDraggablePoints(last);
+                }
+            }
+        }
+    }
+}
